requirements.txt
kafka-python==2.0.2
neo4j==5.9.0
setuptools==68.0.0
wheel==0.40.0